files/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "files-api-server"
version = "0.1.0"
description = "FastAPI server for Elastic Path Files API"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "python-multipart>=0.0.6",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "requests>=2.31.0",
    "pyyaml>=6.0.1"
]

# Add this section to specify package location
[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.292",
    "httpx>=0.27.0",
    "requests>=2.32.3"
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"

